refactor: move visit_relation_expr_opt_alias method from delete.rs to statement.rs
